Add partial containerization for Impinj Interrogators
This commit adds containerization for the Impinj R1000 and R420
interrogators.
The commit creates a Python2 environment that can run the interrogator
scripts.
This commit was completed outside of the Drexel Wireless Systems Lab,
so no testing was using actual impinj hardware, thus the 'partial' in
the commit message.
The client was also not given any exposed ports, as hooking up requests
to the containerized rssi_db container will occur in a separate commit.